refactor: remove indicateConditionValues
