+++ packages/s2-core/__tests__/spreadsheet/multiple-values-cell-spec.tsx
@@ -85,7 +85,6 @@ const getOptions = (): S2Options => {
     showSeriesNumber: true,
     freezeRowHeader: false,
     mode: 'pivot',
-    indicateConditionValues: ['ac', 'rc'],
     conditions: {
       text: [],
       interval: [
