--- packages/s2-core/__tests__/unit/utils/condition/generate-conditions-spec.ts
@@ -1,152 +1,0 @@
-import { INDICATE_CONDITION_SYMBOL } from '@/common/constant/condition';
-import { IconTheme } from '@/common/interface/theme';
-import {
-  getIconPositionCfg,
-  isPositive,
-  updateConditionsByValues,
-} from '@/utils/condition/generate-condition';
-
-describe('Condition Test', () => {
-  const iconTheme: IconTheme = {
-    upIconColor: 'red',
-    downIconColor: 'green',
-  };
-
-  describe('getIconLayoutPosition Test', () => {
-    test('should return right by default', () => {
-      expect(
-        getIconPositionCfg({
-          field: 'value',
-          mapping: () => ({ fill: 'red' }),
-        }),
-      ).toEqual('right');
-    });
-
-    test(`should return left when it's left`, () => {
-      expect(
-        getIconPositionCfg({
-          field: 'value',
-          position: 'left',
-          mapping: () => ({ fill: 'red' }),
-        }),
-      ).toEqual('left');
-    });
-  });
-
-  describe('isPositive Test', () => {
-    test('should expected result when value is number', () => {
-      expect(isPositive(20)).toBeTruthy();
-      expect(isPositive(-20)).toBeFalsy();
-    });
-
-    test('should return expected result when value is string', () => {
-      expect(isPositive('20')).toBeTruthy();
-      expect(isPositive('-20')).toBeFalsy();
-    });
-  });
-
-  describe('updateConditionsByValues Test', () => {
-    test('should return default conditions when there is not values', () => {
-      expect(updateConditionsByValues({}, [], iconTheme)).toEqual({
-        background: [],
-        icon: [],
-        interval: [],
-        text: [],
-      });
-    });
-
-    test('should return expected conditions when there is value without conflict', () => {
-      expect(updateConditionsByValues({}, ['value'], iconTheme)).toEqual({
-        background: [],
-        interval: [],
-        icon: [
-          {
-            type: INDICATE_CONDITION_SYMBOL,
-            field: 'value',
-            position: 'left',
-            mapping: expect.any(Function),
-          },
-        ],
-        text: [
-          {
-            type: INDICATE_CONDITION_SYMBOL,
-            field: 'value',
-            mapping: expect.any(Function),
-          },
-        ],
-      });
-    });
-
-    test('should return expected conditions when there is value with conflict', () => {
-      expect(
-        updateConditionsByValues(
-          {
-            icon: [
-              {
-                field: 'value',
-                position: 'right',
-                mapping: () => ({ fill: 'red' }),
-              },
-            ],
-          },
-          ['value'],
-          iconTheme,
-        ),
-      ).toEqual({
-        background: [],
-        interval: [],
-        icon: [
-          {
-            field: 'value',
-            position: 'right',
-            mapping: expect.any(Function),
-          },
-        ],
-        text: [
-          {
-            type: INDICATE_CONDITION_SYMBOL,
-            field: 'value',
-            mapping: expect.any(Function),
-          },
-        ],
-      });
-    });
-
-    test('should remove generated conditions before update new conditions', () => {
-      expect(
-        updateConditionsByValues(
-          {
-            icon: [
-              {
-                type: INDICATE_CONDITION_SYMBOL,
-                field: 'value',
-                position: 'right',
-                mapping: () => ({ fill: 'red' }),
-              },
-            ],
-          },
-          ['price'],
-          iconTheme,
-        ),
-      ).toEqual({
-        background: [],
-        interval: [],
-        icon: [
-          {
-            type: INDICATE_CONDITION_SYMBOL,
-            field: 'price',
-            position: 'left',
-            mapping: expect.any(Function),
-          },
-        ],
-        text: [
-          {
-            type: INDICATE_CONDITION_SYMBOL,
-            field: 'price',
-            mapping: expect.any(Function),
-          },
-        ],
-      });
-    });
-  });
-});
