+++ packages/s2-core/src/common/constant/condition.ts
@@ -1,3 +1,2 @@
 export const VALUE_RANGES_KEY = 'valueRanges';
 export const DEFAULT_VALUE_RANGES = {};
-export const INDICATE_CONDITION_SYMBOL = Symbol.for('indicate condition'); // used for distinguish whether the condition is generated by indicateConditionValues
