+++ packages/s2-core/src/common/interface/condition.ts
@@ -25,8 +25,6 @@ export type MappingFunction = (
  * One field can hold a condition
  */
 export interface Condition {
-  // internal use it to distinguish whether the condition is generated by indicateConditionValues
-  readonly type?: symbol;
   readonly field: string;
   readonly mapping: MappingFunction;
 }
