+++ packages/s2-core/src/common/interface/s2Options.ts
@@ -39,8 +39,6 @@ export interface S2PartialOptions {
   readonly debug?: boolean;
   // row header hierarchy type only work in pivot mode
   readonly hierarchyType?: 'grid' | 'tree' | 'customTree';
-  // 兜底以前的衍生指标概念
-  readonly indicateConditionValues?: string[];
   // conditions config
   readonly conditions?: Conditions;
   // total config
